Working intel sponza scene GLTF, though takes 40 sec to load. Created a benchmark that shows the bottleneck, will be working on loading optimization next.

+++ Vulkan_Engine/src/EngineCore/VertexBinding.h
@@ -4,7 +4,7 @@
 
 struct MeshDescriptor
 {
-	using TVertexIndices = uint16_t;
+	using TVertexIndices = uint32_t;
 
 	using TVertexPosition = glm::vec3;
 	using TVertexUV = glm::vec2;
